Updates after CR - Added additional file to store auth exceptions - Deleted retries for auth requests to avoid ddos vulnerability - Added aduience parameter to auth redister and verify token methods
