+++ .example_env
@@ -2,3 +2,4 @@ SECRET_KEY=""
 AUTH0_CLIENT_ID=""
 AUTH0_CLIENT_SECRET=""
 AUTH0_DOMAIN=""
+AUTH0_AUDIENCE=""
