+++ app/auth/auth_exceptions.py
@@ -1,0 +1,18 @@
+class TokenVerifierException(Exception):
+    pass
+
+
+class OAuthManagerException(Exception):
+    pass
+
+
+class OAuthServiceUnavailableException(OAuthManagerException):
+    pass
+
+
+class TokenExpiredException(OAuthManagerException):
+    pass
+
+
+class TokenMissingException(OAuthManagerException):
+    pass
