+++ app/config.py
@@ -8,6 +8,7 @@ class Settings(BaseSettings):
     auth0_client_id: str
     auth0_client_secret: str
     auth0_domain: str
+    auth0_audience: str
 
     model_config = SettingsConfigDict(env_file="../.env")
