+++ requirements.txt
@@ -4,4 +4,3 @@ fastapi[standard]
 itsdangerous~=2.2.0
 pydantic-settings~=2.7.0
 httpx~=0.28.1
-tenacity~=9.0.0
